+++ src/main/java/uk/gov/pay/directdebit/payments/services/GoCardlessEventService.java
@@ -2,7 +2,7 @@ package uk.gov.pay.directdebit.payments.services;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.gov.pay.directdebit.payments.dao.GoCardlessEventDao;
+import uk.gov.pay.directdebit.events.dao.GoCardlessEventDao;
 import uk.gov.pay.directdebit.payments.model.GoCardlessEvent;
 
 import javax.inject.Inject;
