+++ src/test/java/uk/gov/pay/directdebit/payments/services/GoCardlessEventServiceTest.java
@@ -7,7 +7,7 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.pay.directdebit.payments.dao.GoCardlessEventDao;
+import uk.gov.pay.directdebit.events.dao.GoCardlessEventDao;
 import uk.gov.pay.directdebit.payments.model.GoCardlessEvent;
 
 import static org.mockito.Mockito.verify;
